Only render controller is requested directly
When requesting a view directly e.g. "src/views/aggregates.php"
the corresponding controller will be instanced, and instance method render
will be called.

If requested by other file (for example inside a route) it won't render, so the route
can call proper methods and manage the output in the response object.